Extracted LocalConnectionAdapterTest units into ConnectionAdapterTestCase
